pyproject.toml
[project]
name = "phase-1-2"
version = "0.1.0"
description = "Mithra - Empathetic AI Companion for Seniors"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-groq>=0.1.0",
    "python-dotenv>=1.0.0",
]
